refactor(frontend): factor out "Item" definitions
Sprinkled throughout the frontend code are
definitions of a mui/Paper theme called "Item".
We copied and pasted it around as we built out
various components. This commit factors it out
to its own component which is then imported
wherever used.

Note: wherever "Item" was replaced with the
      factored out version, I formatted the file.
