+++ src/components/PaperPane/PaperPane.jsx
@@ -1,0 +1,12 @@
+import Paper from '@mui/material/Paper';
+import { styled } from '@mui/material/styles';
+
+const PaperPane = styled(Paper)(({ theme }) => ({
+	backgroundColor: theme.palette.mode === 'dark' ? '#1A2027' : '#fff',
+	...theme.typography.body2,
+	padding: theme.spacing(1),
+	textAlign: 'center',
+	color: theme.palette.text.secondary,
+}));
+
+export { PaperPane }
