+++ src/components/Search/index.js
@@ -1,1 +1,1 @@
-export * from './Observe'
+export * from './Search'
