+++ src/pages/contact/contact.jsx
@@ -1,13 +1,13 @@
-import { Item } from "../observation"
+import { PaperPane } from '../../components/PaperPane/PaperPane'
 
-function Contact(){
-    return (
-        <Item>
-            <h1>Contact Us!</h1>
-            <p>If you notice any issues with this service, you can submit an issue with us <a target="_blank" rel="noreferrer" href="https://github.com/Empyrean-Capstone/Empyrean/issues">here!</a></p>
-            <p>Or, you can contact us at (Insert contact info here)</p>
-        </Item>
-    )
+function Contact() {
+	return (
+		<PaperPane>
+			<h1>Contact Us!</h1>
+			<p>If you notice any issues with this service, you can submit an issue with us <a target="_blank" rel="noreferrer" href="https://github.com/Empyrean-Capstone/Empyrean/issues">here!</a></p>
+			<p>Or, you can contact us at (Insert contact info here)</p>
+		</PaperPane>
+	)
 }
 
-export {Contact}
+export { Contact }
