+++ src/pages/login/login.jsx
@@ -2,32 +2,23 @@ import React from 'react';
 import axios from 'axios';
 
 import './style.css'
+import { PaperPane } from '../../components/PaperPane';
 
 import Grid from '@mui/material/Unstable_Grid2'
 import IconButton from '@mui/material/IconButton';
 import InputAdornment from '@mui/material/InputAdornment';
 import MuiAlert from '@mui/material/Alert';
-import Paper from '@mui/material/Paper';
 import Snackbar from '@mui/material/Snackbar';
 import Stack from '@mui/material/Stack';
 import TextField from '@mui/material/TextField';
 import Visibility from '@mui/icons-material/Visibility';
 import VisibilityOff from '@mui/icons-material/VisibilityOff';
 import { LoadingButton } from "@mui/lab"
-import { styled } from '@mui/material/styles';
 import { useNavigate } from "react-router-dom";
 
 const Alert = React.forwardRef(function Alert(props, ref) {
 	return <MuiAlert elevation={6} ref={ref} variant="filled" {...props} />;
 });
-
-const Item = styled(Paper)(({ theme }) => ({
-	backgroundColor: theme.palette.mode === 'dark' ? '#1A2027' : '#fff',
-	...theme.typography.body2,
-	padding: theme.spacing(1),
-	textAlign: 'center',
-	color: theme.palette.text.secondary,
-}));
 
 
 function Login() {
@@ -65,7 +56,7 @@ function Login() {
 			}
 		>
 			<Grid item>
-				<Item
+				<PaperPane
 					sx={{
 						width: 400,
 						height: '20%',
@@ -177,7 +168,7 @@ function Login() {
 						</Alert>
 					</Snackbar>
 
-				</Item >
+				</PaperPane >
 			</Grid>
 		</Grid >
 	)
